Fin etape 5

+++ ProjetAstronomy/src/ch/epfl/rigel/astronomy/Star.java
@@ -15,7 +15,7 @@ public final class Star extends CelestialObject{
         return hipparcosId;
     }
     
-    public float colorTemperature() {
+    public int colorTemperature() {
         return colorIndex;
     }
